+++ lib/feed/utils/post.dart
@@ -1,18 +1,16 @@
 import 'package:flutter/foundation.dart';
+
 import 'package:lemmy_api_client/v3.dart';
 
 import 'package:thunder/account/account.dart';
 import 'package:thunder/core/enums/enums.dart';
 import 'package:thunder/core/enums/local_settings.dart';
 import 'package:thunder/core/enums/post_sort_type.dart';
-import 'package:thunder/core/enums/subscription_status.dart';
 import 'package:thunder/core/models/models.dart';
 import 'package:thunder/core/singletons/lemmy_client.dart';
 import 'package:thunder/core/singletons/preferences.dart';
 import 'package:thunder/feed/enums/feed_type_subview.dart';
-import 'package:thunder/localizations/app_localizations.dart';
 import 'package:thunder/post/utils/post.dart';
-import 'package:thunder/utils/global_context.dart';
 
 /// Helper function which handles the logic of fetching items for the feed from the API
 /// This includes posts and user information (posts/comments)
@@ -128,142 +126,4 @@ Future<Map<String, dynamic>> fetchFeedItems({
   }
 
   return {'posts': posts, 'comments': comments, 'hasReachedPostsEnd': hasReachedPostsEnd, 'hasReachedCommentsEnd': hasReachedCommentsEnd, 'currentPage': currentPage};
-}
-
-/// Logic to create a post
-Future<PostView> createPost({
-  required int communityId,
-  required String name,
-  String? body,
-  String? url,
-  String? customThumbnail,
-  String? altText,
-  bool? nsfw,
-  int? postIdBeingEdited,
-  int? languageId,
-}) async {
-  final l10n = AppLocalizations.of(GlobalContext.context)!;
-  final account = await fetchActiveProfile();
-  if (account.anonymous) throw Exception(l10n.userNotLoggedIn);
-
-  LemmyApiV3 lemmy = LemmyClient.instance.lemmyApiV3;
-
-  PostResponse postResponse;
-  if (postIdBeingEdited != null) {
-    postResponse = await lemmy.run(EditPost(
-      auth: account.jwt!,
-      name: name,
-      body: body,
-      url: url?.isEmpty == true ? null : url,
-      customThumbnail: customThumbnail?.isEmpty == true ? null : customThumbnail,
-      altText: altText?.isEmpty == true ? null : altText,
-      nsfw: nsfw,
-      postId: postIdBeingEdited,
-      languageId: languageId,
-    ));
-  } else {
-    postResponse = await lemmy.run(CreatePost(
-      auth: account.jwt!,
-      communityId: communityId,
-      name: name,
-      body: body,
-      url: url?.isEmpty == true ? null : url,
-      customThumbnail: customThumbnail?.isEmpty == true ? null : customThumbnail,
-      altText: altText?.isEmpty == true ? null : altText,
-      nsfw: nsfw,
-      languageId: languageId,
-    ));
-  }
-
-  return postResponse.postView;
-}
-
-/// Creates a placeholder post from the given parameters. This is mainly used to display a preview of the post
-/// with the applied settings on Settings -> Appearance -> Posts page.
-Future<ThunderPost?> createExamplePost({
-  String? postTitle,
-  String? postUrl,
-  String? postBody,
-  String? postThumbnailUrl,
-  String? postAltText,
-  bool? locked,
-  bool? nsfw,
-  bool? pinned,
-  String? personName,
-  String? personDisplayName,
-  String? personInstance,
-  String? communityName,
-  String? instanceUrl,
-  int? commentCount,
-  int? scoreCount,
-  bool? saved,
-  bool? read,
-}) async {
-  PostView postView = PostView(
-    post: Post(
-      id: 1,
-      name: postTitle ?? 'Example Title',
-      url: postUrl,
-      body: postBody,
-      thumbnailUrl: postThumbnailUrl,
-      altText: postAltText,
-      creatorId: 1,
-      communityId: 1,
-      removed: false,
-      locked: locked ?? false,
-      published: DateTime.now(),
-      deleted: false,
-      nsfw: nsfw ?? false,
-      apId: '',
-      local: false,
-      languageId: 0,
-      featuredCommunity: pinned ?? false,
-      featuredLocal: false,
-    ),
-    creator: Person(
-      id: 1,
-      name: personName ?? 'Example Username',
-      displayName: personDisplayName ?? 'Example Name',
-      banned: false,
-      published: DateTime.now(),
-      actorId: 'https://$personInstance/u/$personName',
-      local: false,
-      deleted: false,
-      botAccount: false,
-      instanceId: 1,
-    ),
-    community: Community(
-      id: 1,
-      name: communityName ?? 'Example Community',
-      title: '',
-      removed: false,
-      published: DateTime.now(),
-      deleted: false,
-      nsfw: false,
-      actorId: instanceUrl ?? 'https://thunder.lemmy',
-      local: false,
-      hidden: false,
-      postingRestrictedToMods: false,
-      instanceId: 1,
-    ),
-    creatorBannedFromCommunity: false,
-    counts: PostAggregates(
-      id: 1,
-      postId: 1,
-      comments: commentCount ?? 0,
-      score: scoreCount ?? 0,
-      upvotes: 0,
-      downvotes: 0,
-      published: DateTime.now(),
-    ),
-    subscribed: SubscriptionStatus.notSubscribed.toLemmyType(),
-    saved: saved ?? false,
-    read: read ?? false,
-    creatorBlocked: false,
-    unreadComments: 0,
-  );
-
-  List<ThunderPost> posts = await parsePosts([postView]);
-
-  return Future.value(posts.firstOrNull);
 }
